requirements.txt
flask>=2.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
